backend/pyproject.toml
[project]
name = "vibe-margin-backend"
version = "0.1.0"
description = "Recipe API backend with FastAPI and Peewee"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "peewee>=3.17.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "python-multipart>=0.0.20",
  "pydantic>=2.8.0",
  "email-validator>=2.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"
